requirements.txt
agate==1.9.1
annotated-types==0.7.0
attrs==25.3.0
babel==2.17.0
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
clickhouse-connect==0.8.18
colorama==0.4.6
daff==1.4.2
dbt-adapters==1.15.1
dbt-common==1.24.0
dbt-core==1.9.4
dbt-extractor==0.6.0
dbt-postgres==1.9.0
dbt-protos==1.0.291
dbt-semantic-interfaces==0.7.4
deepdiff==7.0.1
greenlet==3.2.2
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kaggle==1.7.4.5
leather==0.4.0
lz4==4.4.4
MarkupSafe==3.0.2
mashumaro==3.14
more-itertools==10.7.0
msgpack==1.1.0
networkx==3.4.2
ordered-set==4.1.0
packaging==25.0
parsedatetime==2.6
pathspec==0.12.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.0
setuptools==80.7.1
six==1.17.0
snowplow-tracker==1.1.0
SQLAlchemy==1.4.54
sqlparse==0.5.3
text-unidecode==1.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
zipp==3.21.0
zstandard==0.23.0
